+++ Tests/Fixtures/simple.php
@@ -1,0 +1,9 @@
+<?php
+/* @var $container \Symfony\Component\DependencyInjection\ContainerBuilder */
+
+$container->loadFromExtension(
+    'simple',
+    array(
+        'types' => array('php')
+    )
+);